requirements.txt
stravalib==0.10.4
google-auth==2.38.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
